core: rename from package gatekeeper
