+++ core/round_tripper.go
@@ -1,4 +1,4 @@
-package gatekeeper
+package core
 
 import (
 	"net/http"
